Decouple feedbacksession and deadline extension update requests
