+++ src/main/java/teammates/ui/constants/ResourceEndpoints.java
@@ -34,6 +34,7 @@ public enum ResourceEndpoints {
     STUDENT_KEY(ResourceURIs.STUDENT_KEY),
     SESSIONS_ONGOING(ResourceURIs.SESSIONS_ONGOING),
     SESSION(ResourceURIs.SESSION),
+    SESSION_DEADLINE_EXTENSIONS(ResourceURIs.SESSION_DEADLINE_EXTENSIONS),
     SESSION_LOGS(ResourceURIs.SESSION_LOGS),
     SESSION_PUBLISH(ResourceURIs.SESSION_PUBLISH),
     SESSION_REMIND_SUBMISSION(ResourceURIs.SESSION_REMIND_SUBMISSION),
