+++ src/web/app/pages-instructor/instructor-session-individual-extension-page/instructor-session-individual-extension-page.component.ts
@@ -26,7 +26,7 @@ import {
 } from '../../../types/api-output';
 import {
   FeedbackSessionBasicRequest,
-  FeedbackSessionUpdateRequest,
+  FeedbackSessionDeadlineExtensionsUpdateRequest,
   Intent,
   ResponseVisibleSetting,
   SessionVisibleSetting,
@@ -374,12 +374,8 @@ export class InstructorSessionIndividualExtensionPageComponent implements OnInit
     isNotifyDeadlines: boolean,
     extensionTimestamp: number,
   ): void {
-    const updatedDeadlinesForCreation = this.getUpdatedDeadlinesForCreation(
+    const request = this.getUpdatedDeadlinesForCreation(
       selectedStudents, selectedInstructors, extensionTimestamp);
-    const request: FeedbackSessionUpdateRequest = {
-      ...updatedDeadlinesForCreation,
-      ...this.feedbackSessionDetails,
-    };
 
     this.handleUpdateDeadlines(request, selectedStudents.length,
       selectedInstructors.length, isNotifyDeadlines, 'created');
@@ -390,18 +386,14 @@ export class InstructorSessionIndividualExtensionPageComponent implements OnInit
     selectedInstructors: InstructorExtensionTableColumnModel[],
     isNotifyDeadlines: boolean,
   ): void {
-    const updatedDeadlinesForDeletion = this.getUpdatedDeadlinesForDeletion(
+    const request = this.getUpdatedDeadlinesForDeletion(
       selectedStudents, selectedInstructors);
-    const request: FeedbackSessionUpdateRequest = {
-      ...updatedDeadlinesForDeletion,
-      ...this.feedbackSessionDetails,
-    };
     this.handleUpdateDeadlines(request, selectedStudents.length,
       selectedInstructors.length, isNotifyDeadlines, 'deleted');
   }
 
   private handleUpdateDeadlines(
-    request: FeedbackSessionUpdateRequest,
+    request: FeedbackSessionDeadlineExtensionsUpdateRequest,
     numStudentsUpdated: number,
     numInstructorsUpdated: number,
     isNotifyDeadlines: boolean,
@@ -409,7 +401,7 @@ export class InstructorSessionIndividualExtensionPageComponent implements OnInit
   ): void {
     this.isSubmittingDeadlines = true;
     this.feedbackSessionsService
-      .updateFeedbackSession(this.courseId, this.feedbackSessionName, request, isNotifyDeadlines)
+      .updateFeedbackSessionDeadlineExtensions(this.courseId, this.feedbackSessionName, request, isNotifyDeadlines)
       .pipe(finalize(() => { this.isSubmittingDeadlines = false; }))
       .subscribe({
         next: () => {
