+++ src/web/services/feedback-sessions.service.ts
@@ -27,6 +27,7 @@ import {
 } from '../types/api-output';
 import {
   FeedbackSessionCreateRequest,
+  FeedbackSessionDeadlineExtensionsUpdateRequest,
   FeedbackSessionRespondentRemindRequest,
   FeedbackSessionUpdateRequest,
   Intent,
@@ -104,15 +105,27 @@ export class FeedbackSessionsService {
   /**
    * Updates a feedback session by calling API.
    */
-  updateFeedbackSession(courseId: string, feedbackSessionName: string, request: FeedbackSessionUpdateRequest,
-    isNotifyDeadlines: boolean = false):
+  updateFeedbackSession(courseId: string, feedbackSessionName: string, request: FeedbackSessionUpdateRequest):
       Observable<FeedbackSession> {
+    const paramMap: Record<string, string> = {
+      courseid: courseId,
+      fsname: feedbackSessionName,
+    };
+    return this.httpRequestService.put(ResourceEndpoints.SESSION, paramMap, request);
+  }
+
+  /**
+   * Updates the deadline extensions for a feedback session by calling API.
+   */
+  updateFeedbackSessionDeadlineExtensions(courseId: string, feedbackSessionName: string,
+    request: FeedbackSessionDeadlineExtensionsUpdateRequest,
+    isNotifyDeadlines: boolean = false): Observable<FeedbackSessionDeadlineExtensions> {
     const paramMap: Record<string, string> = {
       courseid: courseId,
       fsname: feedbackSessionName,
       notifydeadlines: String(isNotifyDeadlines),
     };
-    return this.httpRequestService.put(ResourceEndpoints.SESSION, paramMap, request);
+    return this.httpRequestService.put(ResourceEndpoints.SESSION_DEADLINE_EXTENSIONS, paramMap, request);
   }
 
   /**
